+++ source/ReadAlign_outputAlignments.cpp
@@ -32,7 +32,7 @@ void ReadAlign::outputAlignments() {
         };
 
         //the operations below are both for mapped and unmapped reads
-        soloRead->readBar->getCBandUMI(readBarcodeSeq, readBarcodeQual, readNameExtra[0], readFilesIndex, readName);
+        soloRead->readBar->getCBandUMI(Read0, Qual0, readLengthOriginal, readNameExtra[0], readFilesIndex, readName);
 
         soloRead->record((unmapType<0 ? nTr : 0), trMult, iReadAll, readAnnot); //need to supply nTr=0 for unmapped reads